# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Learnup",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/notifications/page": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get paginated notifications",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/types.PaginationRes"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "row": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/github_com_ladmakhi81_learnup_internals_notification_dto_res.NotificationPageItem"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    }
                }
            }
        },
        "/notifications/{notification-id}/seen": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Mark notification as seen",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notification ID",
                        "name": "notification-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    }
                }
            }
        },
...
        "/payments/page": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get paginated payments",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/types.PaginationRes"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "rows": {
                                                            "$ref": "#/definitions/github_com_ladmakhi81_learnup_internals_payment_dto_res.GetPageablePaymentItem"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    }
                }
            }
        },
...
        "/teacher/comments/{course-id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teacher"
                ],
                "summary": "Get paginated comments by course ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "course-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/types.PaginationRes"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "rows": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/github_com_ladmakhi81_learnup_internals_teacher_dto_res.GetCommentPageableItemRes"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.PaymentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "success",
                "failed"
            ],
            "x-enum-varnames": [
                "PaymentStatus_Pending",
                "PaymentStatus_Success",
                "PaymentStatus_Failure"
            ]
        },
...
        "github_com_ladmakhi81_learnup_internals_notification_dto_res.NotificationPageItem": {
            "type": "object",
            "properties": {
                "isSeen": {
                    "type": "boolean"
                },
                "metadata": {},
                "seenAt": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_ladmakhi81_learnup_internals_notification_dto_res.userItem"
                }
            }
        },
        "github_com_ladmakhi81_learnup_internals_notification_dto_res.userItem": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_ladmakhi81_learnup_internals_payment_dto_res.GetPageablePaymentItem": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "gateway": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/entities.PaymentStatus"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_ladmakhi81_learnup_internals_teacher_dto_res.GetCommentPageableItemRes": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "course": {
                    "$ref": "#/definitions/github_com_ladmakhi81_learnup_internals_teacher_dto_res.getCommentCourseItem"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_ladmakhi81_learnup_internals_teacher_dto_res.getCommentUserItem"
                }
            }
        },
...
        "github_com_ladmakhi81_learnup_internals_teacher_dto_res.getCommentCourseItem": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_ladmakhi81_learnup_internals_teacher_dto_res.getCommentUserItem": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "types.ApiError": {
            "type": "object",
            "properties": {
                "message": {},
                "statusCode": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "integer"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
        "types.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "types.PaginationRes": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "row": {},
                "totalCount": {
                    "type": "integer"
                },
                "totalPage": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}